Update tests and get all tests passing
